Hurricane Ida
Tableau: For BI reporting I am using Tableau to visualize data in charts, graphs, or histograms and presenting insights 

Hurricane Ida was a deadly Atlantic hurricane in 2021. I created a Dashboard to assess the impact of Ida on banking industry ,on deposit growth and deposit presence for those specific institutions that located in the hurricane Ida's path. Because deposit growth was a good factor that showed customer trust the institution and showed institution's health.